tap-api/requirements.txt
aiodocker==0.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
alembic==1.16.1
anyio==4.9.0
arrow==1.3.0
atomicwrites==1.4.1
attrs==25.3.0
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
check-jsonschema==0.33.0
click==8.2.1
click-default-group==1.2.4
click-didyoumean==0.3.1
cookiecutter==2.6.0
croniter==6.0.0
dateparser==1.2.1
distlib==0.3.9
fasteners==0.19
filelock==3.18.0
flatten-dict==0.4.2
frozenlist==1.6.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meltano==3.7.8
multidict==6.4.4
packaging==25.0
platformdirs==4.3.8
propcache==0.3.1
psutil==7.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
regress==2025.5.1
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
structlog==25.4.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
uv==0.7.11
virtualenv==20.31.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
